Commented out ``install_requires`` in setup.py.
If ``install_requires`` is used, then, when the module was installed in a virtual environment using ``pip``, it was not checked whether a dependency already existed or not, but it was always re-installed. This caused some trouble with modules, like PIL, which require some development libraries in order to have full support for various image formats.